backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
dj-rest-auth==5.0.2
drf-spectacular==0.26.5
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Pillow==10.1.0
